libs/workflow_engine/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ztc-workflow-engine"
version = "0.1.0"
description = "ZTC Workflow Engine - Core workflow orchestration and execution"
readme = "README.md"
license = "MIT"
authors = ["ZeroTouch Team"]
keywords = ["workflow", "orchestration", "kubernetes", "gitops", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "workflow_engine", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
pydantic = ">=2.0.0,<3.0.0"
jinja2 = ">=3.0.0,<4.0.0"
pyyaml = ">=6.0.0,<7.0.0"
ruamel-yaml = ">=0.18.0,<0.19.0"
aiohttp = ">=3.9.0,<4.0.0"
cryptography = ">=42.0.0,<43.0.0"
requests = ">=2.31.0,<3.0.0"
pyjwt = ">=2.8.0,<3.0.0"
aiofiles = ">=24.0.0,<25.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0,<9.0.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
hypothesis = ">=6.0.0,<7.0.0"

[tool.poetry.urls]
Repository = "https://github.com/zerotouch/ztc"
Issues = "https://github.com/zerotouch/ztc/issues"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/workflow_engine"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Tests expect to run from zerotouch-engine root (where platform/platform.yaml exists)
# Use: cd ../.. && python -m pytest libs/workflow_engine/src/workflow_engine/adapters/
